requirements.txt
streamlit==1.28.0
mysql-connector-python==8.2.0
numpy==1.24.3
pandas==2.0.3
Pillow==10.1.0
python-dotenv==1.0.0
streamlit-authenticator==0.2.3
plotly==5.17.0
bcrypt==4.0.1
